# Conflicts:
#	docker/docker-compose-prod.yml
#	frontend/src/app/component/sidebar/sidebar.component.html

+++ backend/src/main/kotlin/com/mr/websocket_chat/service/ChatRoomService.kt
@@ -1,48 +1,71 @@
 package com.mr.websocket_chat.service
 
+import com.mr.websocket_chat.domain.exception.ChatRoomNotFoundException
+import com.mr.websocket_chat.domain.exception.UserNotFoundException
 import com.mr.websocket_chat.domain.jpa.ChatRoomEntity
+import com.mr.websocket_chat.domain.mapper.ChatRoomMapper
+import com.mr.websocket_chat.domain.rest.ChatRoomDTO
 import com.mr.websocket_chat.repository.ChatRoomRepository
+import com.mr.websocket_chat.repository.UserRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import kotlin.jvm.optionals.getOrNull
 
 @Service
 class ChatRoomService @Autowired constructor(
-	private val chatRoomRepository: ChatRoomRepository
+	private val chatRoomRepository: ChatRoomRepository,
+	private val userRepository: UserRepository,
+	private val chatRoomMapper: ChatRoomMapper
 ){
 
-	fun addChatRoom(room: ChatRoomEntity): ChatRoomEntity {
-		return chatRoomRepository.save(room)
+	@Transactional
+	fun createChatRoom(roomName: String, creator: String): ChatRoomDTO {
+		val creatorEntity = userRepository.findByUsername(creator) ?: throw UserNotFoundException()
+		val createdRoom = chatRoomRepository.save(
+			ChatRoomEntity(
+				roomName,
+				mutableSetOf(creatorEntity)
+			)
+		)
+		return chatRoomMapper.toDTO(createdRoom)
 	}
 
-	fun findById(roomId: Long): ChatRoomEntity? {
-		return chatRoomRepository.findByIdOrNull(roomId)
+	@Transactional(readOnly = true)
+	fun getChatRoomsForUser(username: String): List<ChatRoomDTO> {
+		return chatRoomRepository.findByUsers_Username(username).map {
+			chatRoomMapper.toDTO(it)
+		}
 	}
 
-	fun getChatRoomsForUser(username: String): List<ChatRoomEntity> {
-		return chatRoomRepository.findByUsers_Username(username)
-	}
-
-	fun removeUserFromRoom(username: String, roomId: Long): Boolean {
-		val roomEntity = chatRoomRepository.findByIdOrNull(roomId) ?: return false
-		roomEntity.users.removeIf { element -> element.username == username }
+	@Transactional
+	fun removeUserFromRoom(roomId: Long, username: String) {
+		val roomEntity = chatRoomRepository.findByIdOrNull(roomId) ?: throw ChatRoomNotFoundException()
+		val wasAnyoneRemoved = roomEntity.users.removeIf { element -> element.username == username }
 		if(roomEntity.users.isEmpty()) {
 			chatRoomRepository.deleteById(roomId)
-		} else {
+		} else if(wasAnyoneRemoved) {
 			chatRoomRepository.save(roomEntity)
 		}
-		return true
 	}
 
+	@Transactional(readOnly = true)
 	fun chatRoomExists(roomId: Long): Boolean {
 		return chatRoomRepository.existsById(roomId)
 	}
 
-	fun deleteChatRoom(roomId: Long) {
-		chatRoomRepository.deleteById(roomId)
+	@Transactional
+	fun deleteChatRoom(roomId: Long, deletingUser: String) {
+		val roomEntity = chatRoomRepository.findByIdOrNull(roomId) ?: throw ChatRoomNotFoundException()
+		if(isUserChatRoomMember(deletingUser, roomEntity)) {
+			chatRoomRepository.deleteById(roomId)
+		} else {
+			throw RuntimeException("User cannot delete room to which he does not belong. User id: ${deletingUser}, room id: $roomId.")
+		}
 	}
 
+	@Transactional(readOnly = true)
 	fun isUserChatRoomMember(username: String, roomId: Long): Boolean {
 		val room = chatRoomRepository.findById(roomId).getOrNull() ?: return false
 		return isUserChatRoomMember(username, room)
@@ -50,5 +73,17 @@ class ChatRoomService @Autowired constructor(
 
 	fun isUserChatRoomMember(username: String, room: ChatRoomEntity): Boolean {
 		return room.users.any { it.username == username }
+	}
+
+	@Transactional
+	fun changeRoomName(roomId: Long, newRoomName: String, changingUser: String): ChatRoomDTO {
+		val roomEntity = chatRoomRepository.findByIdOrNull(roomId) ?: throw ChatRoomNotFoundException()
+		if(!isUserChatRoomMember(changingUser, roomEntity)) {
+			throw RuntimeException("User cannot change the name of a room to which he does not belong. User id: ${changingUser}, room id: $roomId.")
+		}
+
+		roomEntity.name = newRoomName
+		chatRoomRepository.save(roomEntity)
+		return chatRoomMapper.toDTO(roomEntity)
 	}
 }
